Make the examples runnable without top-level code
